+++ release-notes.rst
@@ -2,6 +2,12 @@ Release notes for chevah.compat
 ===============================
 
 
+1.5.0 - 2025-03-19
+------------------
+
+* Add API to authenticate OS users with only PAM and without root requirement.
+* Remove `conditionals.onOSVersion` helper.
+
 1.4.0 - 2024-12-08
 ------------------